@@ -1,19 +1,18 @@
 ﻿using Microsoft.Extensions.VectorData;
 
-namespace VectorDataAI
+namespace VectorDataAI;
+
+internal class CloudService
 {
-    internal class CloudService
-    {
-        [VectorStoreRecordKey]
-        public int Key { get; set; }
+    [VectorStoreRecordKey]
+    public int Key { get; set; }
 
-        [VectorStoreRecordData]
-        public string Name { get; set; }
+    [VectorStoreRecordData]
+    public string Name { get; set; }
 
-        [VectorStoreRecordData]
-        public string Description { get; set; }
+    [VectorStoreRecordData]
+    public string Description { get; set; }
 
-        [VectorStoreRecordVector(384, DistanceFunction.CosineSimilarity)]
-        public ReadOnlyMemory<float> Vector { get; set; }
-    }
+    [VectorStoreRecordVector(Dimensions: 384, DistanceFunction = DistanceFunction.CosineSimilarity)]
+    public ReadOnlyMemory<float> Vector { get; set; }
 }
